+++ src/apps/explorer/pages/styled.tsx
@@ -1,6 +1,8 @@
 import styled from 'styled-components'
 import { media } from 'theme/styles/media'
 
+import { RowWithCopyButton } from 'components/common/RowWithCopyButton'
+
 export const WrapperPage = styled.div`
   padding: 1.6rem;
   margin: 0 auto;
@@ -20,5 +22,16 @@ export const WrapperPage = styled.div`
     padding: 2.4rem 0 0.75rem;
     align-items: center;
     font-weight: ${({ theme }): string => theme.fontBold};
+  }
+`
+
+export const TitleAddress = styled(RowWithCopyButton)`
+  font-size: ${({ theme }): string => theme.fontSizeDefault};
+  font-weight: ${({ theme }): string => theme.fontNormal};
+  margin: 0 0 0 1.5rem;
+  display: flex;
+  align-items: center;
+  ${media.tinyDown} {
+    font-size: 1.2rem;
   }
 `
